Prevent wrapping in minimap buffer
When wrap is on and showbreak is enabled the minimap buffer looks terrible. To see what I mean add set wrap and set showbreak=.... (That's four periods or dots) to your vimrc, then open vim and minimap. If it doesn't look terrible, add the following autocmd to your vimrc and restart vim: autocmd! VimEnter,BufRead * Minimap