I moved some recurring code (searching for the minimap window) to its own function
-Made a function to add autocmd or clear the autocmd
-All basic (update, open and close) functions use the new functions
-update function now searches for the minimap window as a start to see
if it's updating itself
-make an extra function (called toggle) which does excatly the same
(check with the new search for minimap window function) if the window is
there and call the close or open function
-added an try/except when closing the window
Plus some minor changes to the .vim files to support the new python
functions.
You will see it's still easy to follow even though it's one commit, all
bug's and the feature depend on the new functions.